Add support for GPG explicit keys
When passing numeric key identifiers to GPG, GPG does not use them
verbatim, and instead converts them into an identity which it then
uses for encryption. This means that GPG can not be told to use all
of the encryption capable subkeys as it will collapse them all into
the same identity, and then select only one of the keys.

To make GPG do the right thing, each of the subkeys must be suffixed
with an exclamation mark (`!`) to tell it to use the key identifier
verbatim and not perform conversion to an identity first. This way
many encryption capable subkeys can be specified in the `.gpg-id` of
a password store.